requirements.txt
python-telegram-bot==21.5
pytz==2024.2